requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0
django-allauth==0.40.0
django-countries==5.5
django-crispy-forms==1.8.1
heroku==0.1.4
idna==2.8
oauthlib==3.1.0
Pillow==6.2.1
pkg-resources==0.0.0
python-dateutil==1.5
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
